utils/lightbar_utils/pyproject.toml
[tool.poetry]
name = "lightbar-utils"
version = "0.1.0"
description = ""
authors = ["Radu_Tolontan <radut@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pillow = "^11.1.0"
numpy = "^2.2.3"
matplotlib = "^3.10.1"
tk = "^0.1.0"
image = "^1.5.33"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
